[MOD] Store Module: Compactify data before storage

+++ basex-core/src/main/java/org/basex/query/value/seq/DblSeq.java
@@ -98,7 +98,7 @@ public final class DblSeq extends NativeSeq {
    * @return value
    * @throws QueryException query exception
    */
-  static Value get(final int size, final Value... values) throws QueryException {
+  static Value get(final long size, final Value... values) throws QueryException {
     final DoubleList tmp = new DoubleList(size);
     for(final Value value : values) {
       // speed up construction, depending on input
